Refactored device attribute handling out of MatrixFlow
