+++ src/module/flows/MatrixFlow.ts
@@ -1,54 +1,11 @@
 import { SR5Actor } from '../actor/SR5Actor';
 import { SR5Item } from '../item/SR5Item';
-import { MatrixAttributesType } from '../types/template/Matrix';
 import { Translation } from '../utils/strings';
 
 /**
  * General handling around everything matrix related.
  */
 export const MatrixFlow = {
-    /**
-     * Switch out one matrix attribute with a new one.
-     *
-     * NOTE: attributes given are changed in place.
-     *
-     * @param matrixAttributes The attribute data.
-     * @param changedSlot The attribute slot changed
-     * @param attribute The attribute selected to change into the changed slot.
-     *
-     * @returns Changed matrix attributes data.
-     */
-    changeMatrixAttribute(
-        matrixAttributes: MatrixAttributesType,
-        changedSlot: string,
-        attribute: Shadowrun.MatrixAttribute,
-    ): Record<string, any> {
-        // The attribute that used to be in the slot that's changing.
-        const previousAttribute = matrixAttributes[changedSlot].att;
-        // The slot of the selected attribute that will get the previous attribute.
-        let previousSlot = '';
-        Object.entries(matrixAttributes).forEach(([slot, { att }]) => {
-            if (att === attribute) {
-                previousSlot = slot;
-            }
-        });
-
-        if (!previousSlot) {
-            console.error(
-                `Shadowrun 5e | Couldn't change attribute ${attribute} as it wasn't found in matrix attribute slots`,
-                matrixAttributes,
-            );
-            return {};
-        }
-
-        const updateData = {
-            [`system.atts.${changedSlot}.att`]: attribute,
-            [`system.atts.${previousSlot}.att`]: previousAttribute,
-        };
-
-        return updateData;
-    },
-
     /**
      * Collect all matrix actions of an actor.
      * @param actor The actor to collect matrix actions from.
@@ -63,7 +20,7 @@ export const MatrixFlow = {
     },
 
     /**
-     * Trasnform the given document to a string type for sheet display.
+     * Transform the given document to a string type for sheet display.
      *
      * NOTE: This function is part of sheet rendering, so we fail silently, to not break sheet rendering.
      *
